Move source files to specific dir and fix imports
