+++ src/kde_material_you_colors/utils/ksyntax_utils.py
@@ -1,6 +1,6 @@
 import json
 import os
-import settings
+from .. import settings
 
 
 def export_schemes(schemes=None):
