+++ src/kde_material_you_colors/utils/pywal_sequences_timeout.py
