+++ src/kde_material_you_colors/utils/utils.py
@@ -2,7 +2,7 @@ import gettext
 import logging
 import os
 import subprocess
-import settings
+from .. import settings
 import argparse
 import sys
 import re
